guard added

+++ src/image/image.service.ts
@@ -1,32 +1,58 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpException,
+  HttpStatus,
+  Inject,
+  Injectable,
+  forwardRef,
+} from '@nestjs/common';
 import { Response } from 'express';
 import { Storage } from '@google-cloud/storage';
 import { extname } from 'path';
+import { InjectModel } from '@nestjs/sequelize';
+import { Image } from './models/image.model';
+import { UserService } from '../user/user.service';
+import { v4 } from 'uuid';
 
 const storage = new Storage({
   projectId: 'upload-image-392818',
   keyFilename: 'keyfile.json',
 });
 
+const bucketName = 'upload-image-nest-blog';
+const bucket = storage.bucket(bucketName);
+
 @Injectable()
 export class ImageService {
-  async create(image: Express.Multer.File) {
-    try {
-      const bucketName = 'upload-image-nest-blog';
-      const bucket = storage.bucket(bucketName);
+  constructor(
+    @InjectModel(Image)
+    private readonly imageRepository: typeof Image,
+    @Inject(forwardRef(() => UserService))
+    private readonly userService: UserService,
+  ) {}
 
+  async create(image: Express.Multer.File, authHeader: string) {
+    if (!image) throw new BadRequestException('No image');
+    const user = await this.userService.verifyToken(authHeader);
+
+    try {
       const fileName =
         (await this.generateUniqueFileName()) + extname(image.originalname);
       const file = bucket.file(fileName);
 
       await file.save(image.buffer, { resumable: false });
+      await this.imageRepository.create({
+        id: v4(),
+        file_name: fileName,
+        user_id: user.id,
+      });
+
       const url = `${process.env.API_BASE_URL}/api/image/${fileName}`;
       return {
         fileName,
         url,
       };
     } catch (error) {
-      console.log(error);
       throw new HttpException(
         'Error with uploading images',
         HttpStatus.INTERNAL_SERVER_ERROR,
@@ -34,21 +60,16 @@ export class ImageService {
     }
   }
 
-  async findAll() {
-    const bucketName = 'upload-image-nest-blog';
-    const bucket = storage.bucket(bucketName);
-
-    const [files] = await bucket.getFiles();
-    const allFileNames = files.map((file) => file.name);
-
-    return allFileNames;
+  async findAll(authHeader: string) {
+    const user = await this.userService.verifyToken(authHeader);
+    return this.imageRepository.findAll({
+      where: { user_id: user.id },
+      attributes: ['id', 'file_name', 'createdAt'],
+    });
   }
 
   async findOne(fileName: string, res: Response) {
-    const bucketName = 'upload-image-nest-blog';
-    const bucket = storage.bucket(bucketName);
     const file = bucket.file(fileName);
-
     const exists = await file.exists();
     if (!exists[0]) {
       throw new HttpException('Image not found', HttpStatus.NOT_FOUND);
@@ -58,17 +79,42 @@ export class ImageService {
     stream.pipe(res);
   }
 
-  async remove(fileName: string) {
-    const bucketName = 'upload-image-nest-blog';
-    const bucket = storage.bucket(bucketName);
-    const file = bucket.file(fileName);
+  async remove(fileName: string, authHeader: string) {
+    const image = await this.imageRepository.findOne({
+      where: { file_name: fileName },
+    });
+    if (!image) {
+      throw new HttpException('Image not found', HttpStatus.NOT_FOUND);
+    }
 
+    await this.userService.isUserSelf(image.user_id, authHeader);
+
+    const file = bucket.file(fileName);
     const exists = await file.exists();
     if (exists[0]) {
       await file.delete();
+      await this.imageRepository.destroy({ where: { file_name: fileName } });
       return fileName;
     } else {
       throw new HttpException('Image not found', HttpStatus.NOT_FOUND);
+    }
+  }
+
+  async removeAllImageByUserId(user_id: string) {
+    try {
+      const images = await this.imageRepository.findAll({
+        where: { user_id },
+        attributes: ['id', 'file_name'],
+      });
+      for (let image of images) {
+        const file = bucket.file(image.file_name);
+        const exists = await file.exists();
+        if (exists[0]) {
+          await file.delete();
+        }
+      }
+    } catch (error) {
+      throw new BadRequestException(error.message);
     }
   }
 
@@ -83,9 +129,6 @@ export class ImageService {
   }
 
   async generateUniqueFileName() {
-    const bucketName = 'upload-image-nest-blog';
-    const bucket = storage.bucket(bucketName);
-
     const [files] = await bucket.getFiles();
     const allUniqueFileNames = files.map((file) => file.name);
